working on #91
