+++ src/main/java/org/woehlke/twitterwall/frontend/controller/entities/HashTagController.java
@@ -10,11 +10,15 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.woehlke.twitterwall.frontend.common.AbstractTwitterwallController;
 import org.woehlke.twitterwall.frontend.common.Symbols;
+import org.woehlke.twitterwall.frontend.model.HashTagCounted;
 import org.woehlke.twitterwall.oodm.entities.Tweet;
 import org.woehlke.twitterwall.oodm.entities.User;
+import org.woehlke.twitterwall.oodm.entities.entities.HashTag;
 import org.woehlke.twitterwall.oodm.service.TweetService;
 import org.woehlke.twitterwall.oodm.service.UserService;
+import org.woehlke.twitterwall.oodm.service.entities.HashTagService;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -63,6 +67,42 @@ public class HashTagController extends AbstractTwitterwallController {
         }
     }
 
+    @RequestMapping("/hashtag/overview")
+    public String hashTags(Model model) {
+        String msg = "/hashtags: ";
+        logEnv();
+        String title = "HashTags";
+        String subtitle = searchterm;
+        String symbol = Symbols.HASHTAG.toString();
+        model = setupPage(model,title,subtitle,symbol);
+        List<HashTagCounted> hashTagsTweets = new ArrayList<>();
+        List<HashTagCounted> hashTagsUsers = new ArrayList<>();
+        for (HashTag hashTag : hashTagService.getAll()) {
+            String text = hashTag.getText();
+            try {
+                long numberTweets = tweetService.countTweetsForHashTag(hashTag.getText());
+                if(numberTweets > 0) {
+                    HashTagCounted c = new HashTagCounted(numberTweets, text);
+                    hashTagsTweets.add(c);
+                }
+            } catch (IllegalArgumentException e){
+                log.warn(msg+"tweetService.countTweetsForHashTag: "+e.getMessage());
+            }
+            try {
+                long numberUsers = userService.countUsersForHashTag(hashTag.getText());
+                if(numberUsers > 0){
+                    HashTagCounted c = new HashTagCounted(numberUsers,text);
+                    hashTagsUsers.add(c);
+                }
+            } catch (IllegalArgumentException e){
+                log.warn(msg+"tweetService.countTweetsForHashTag: "+e.getMessage());
+            }
+        }
+        model.addAttribute("hashTagsTweets", hashTagsTweets);
+        model.addAttribute("hashTagsUsers", hashTagsUsers);
+        return "tags";
+    }
+
     private static final Logger log = LoggerFactory.getLogger(HashTagController.class);
 
     @Value("${twitterwall.frontend.menu.appname}")
@@ -89,12 +129,15 @@ public class HashTagController extends AbstractTwitterwallController {
     @Value("${twitterwall.frontend.idGoogleAnalytics}")
     private String idGoogleAnalytics;
 
+    private final HashTagService hashTagService;
+
     private final TweetService tweetService;
 
     private final UserService userService;
 
     @Autowired
-    public HashTagController(TweetService tweetService, UserService userService) {
+    public HashTagController(HashTagService hashTagService, TweetService tweetService, UserService userService) {
+        this.hashTagService = hashTagService;
         this.tweetService = tweetService;
         this.userService = userService;
     }
