+++ src/main/resources/templates/layout.html
@@ -62,14 +62,14 @@
                             <li><a href="/user/notyetfriends" th:href="@{/user/notyetfriends}"><i class="fa fa-plus-square" aria-hidden="true"></i>&nbsp;&nbsp;Not Yet Friends</a></li>
                             <li><a href="/user/notyetonlist" th:href="@{/user/notyetonlist}"><i class="fa fa-handshake-o" aria-hidden="true"></i>&nbsp;&nbsp;Not Yet On List</a></li>
                             <li><a href="/user/onlist" th:href="@{/user/onlist}"><i class="fa fa-leaf" aria-hidden="true"></i>&nbsp;&nbsp;On List</a></li>
-                            <li><a href="/user/onlist/renew" th:href="@{/user/onlist/renew}"><i class="fa fa-leaf" aria-hidden="true"></i>&nbsp;&nbsp;Renew List of Users On List</a></li>
-                            <li><a href="/getTestData" th:href="@{/getTestData}"><i class="fa fa-cubes" aria-hidden="true"></i>&nbsp;&nbsp;Create Test Data</a></li>
-                            <li><a href="/service/count" th:href="@{/service/count}"><i class="fa fa-database" aria-hidden="true"></i>&nbsp;&nbsp;Service Count</a></li>
+                            <li><a href="/test/user/onlist/renew" th:href="@{/test/user/onlist/renew}"><i class="fa fa-leaf" aria-hidden="true"></i>&nbsp;&nbsp;Renew List of Users On List</a></li>
+                            <li><a href="/test/getTestData" th:href="@{/test/getTestData}"><i class="fa fa-cubes" aria-hidden="true"></i>&nbsp;&nbsp;Create Test Data</a></li>
+                            <li><a href="/application/domain/count" th:href="@{/application/domain/count}"><i class="fa fa-database" aria-hidden="true"></i>&nbsp;&nbsp;Count Bom</a></li>
                             <li><a href="/task/all" th:href="@{/task/all}"><i class="fa fa-check-square" aria-hidden="true"></i>&nbsp;&nbsp;Tasks</a></li>
                         </ul>
                     </li>
                     <li><a href="/user/tweets" th:href="@{/user/tweets}"><i class="fa fa-users" aria-hidden="true"></i>&nbsp;&nbsp;Users</a></li>
-                    <li><a href="/hashtags" th:href="@{/hashtags}"><i class="fa fa-hashtag" aria-hidden="true"></i>&nbsp;&nbsp;HashTags</a></li>
+                    <li><a href="/hashtag/overview" th:href="@{/hashtag/overview}"><i class="fa fa-hashtag" aria-hidden="true"></i>&nbsp;&nbsp;HashTags</a></li>
                     <li><a href="/imprint" th:href="@{/imprint}"><i class="fa fa-university" aria-hidden="true"></i>&nbsp;&nbsp;Impressum</a></li>
                 </ul>
                 <!--
